pyproject.toml
[project]
name = "mobius"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "flet-desktop>=0.80.2",
  "flet[all]>=0.80.4",
  "mokkari>=3.17.0",
  "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
  "flet[all]>=0.80.4",
]

[tool.flet]
org = "com.mycompany"
product = "mobius"
company = "Flet"
copyright = "Copyright (C) 2026 by Flet"

[tool.flet.app]
path = "src"
